requirements.txt
discord.py[audio]
PyNaCl
yt-dlp